user-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
PyJWT==2.8.0
python-dotenv==1.0.0
SQLAlchemy==2.0.23
databases==0.8.0
psycopg2-binary==2.9.9
passlib[bcrypt]==1.7.4
